Fixed an issue with comments in the site list
A comment to a feedback entry can be multine, and the API for the backoffice returns a string array containing each line, so the lines can be iterated. However if two or more lines are equal, ngRepeat fails since it can't tell the string items apart. The fix creates a ghost array where each string item is wrapped in an object, which works fine with ngRepeat.