The "SetStatus" and "SetAssignedTo" methods should update the "FeedbackEntry" instance directly instead of the underlying "FeedbackDatabaseEntry" instance
When only updating the "FeedbackDatabaseEntry" instance, changes won't be reflected in the corresponding "FeedbackEntry" instance. This the status or responsible user before the changes is returned in the API instead of the new value. The code has now been updated to change the "FeedbackEntry" instance directly, which then also automatically updates the "FeedbackDatabaseEntry" instance.